feat: AmplitudeClient 통신 추가
- amplitude rest api의 응답 결과 데이터를 기반으로 반환
- 조회 관련 데이터 integer -> long
- year, startDate, endDate 공통 상수로 이동

related_issue: #581

+++ src/main/java/org/sopt/makers/internal/report/dto/PlayGroundTypeStatsDto.java
@@ -6,16 +6,16 @@ import java.util.List;
 import org.sopt.makers.internal.report.domain.PlaygroundType;
 
 public record PlayGroundTypeStatsDto(
-	Integer community,
-	Integer member,
-	Integer project,
-	Integer wordChainGame,
-	Integer coffeeChat,
-	Integer crew
+	Long community,
+	Long member,
+	Long project,
+	Long wordChainGame,
+	Long coffeeChat,
+	Long crew
 ) {
 	public PlaygroundType getTopStats() {
-		List<Integer> values = Arrays.asList(community, member, project, wordChainGame, coffeeChat, crew);
-		int max = values.stream().max(Integer::compareTo).orElse(0);
+		List<Long> values = Arrays.asList(community, member, project, wordChainGame, coffeeChat, crew);
+		long max = values.stream().max(Long::compareTo).orElse(0L);
 
 		if (max == community) {
 			return PlaygroundType.COMMUNITY;
